Package name refactor + small code cleanup
