+++ app/build.gradle
@@ -13,7 +13,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.surfaceduo.training.itemslist"
+        applicationId "com.surfaceduo.listitems"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
