+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
   ~
   -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.surfaceduo.training.listitems">
+    package="com.surfaceduo.listitems">
 
     <application
         android:allowBackup="true"
@@ -13,7 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar"> <!-- we have to use a theme without action bar -->
-        <activity android:name="com.surfaceduo.training.listitems.MainActivity">
+        <activity android:name="com.surfaceduo.listitems.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
