+++ app/src/main/java/com/surfaceduo/listitems/DetailFragment.kt
@@ -4,7 +4,7 @@
  *
  */
 
-package com.surfaceduo.training.listitems
+package com.surfaceduo.listitems
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -16,7 +16,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
 import com.microsoft.device.dualscreen.core.ScreenHelper
-import com.surfaceduo.training.listitems.recyclerview.Item
+import com.surfaceduo.listitems.recyclerview.Item
 
 class DetailFragment : Fragment() {
