+++ app/src/main/java/com/surfaceduo/listitems/MainActivity.kt
@@ -4,7 +4,7 @@
  *
  */
 
-package com.surfaceduo.training.listitems
+package com.surfaceduo.listitems
 
 import android.os.Bundle
 import android.view.MenuItem
