+++ app/src/main/java/com/surfaceduo/listitems/SharedVM.kt
@@ -4,11 +4,11 @@
  *
  */
 
-package com.surfaceduo.training.listitems
+package com.surfaceduo.listitems
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.surfaceduo.training.listitems.recyclerview.Item
+import com.surfaceduo.listitems.recyclerview.Item
 
 class SharedVM : ViewModel() {
     val selectedItem = MutableLiveData<Item>()
