+++ app/src/main/java/com/surfaceduo/listitems/recyclerview/Item.kt
@@ -1,0 +1,3 @@
+package com.surfaceduo.listitems.recyclerview
+
+data class Item(val number: Int, val body: String)
