--- app/src/main/java/com/surfaceduo/training/listitems/recyclerview/Item.kt
@@ -1,3 +1,0 @@
-package com.surfaceduo.training.listitems.recyclerview
-
-data class Item(val number: Int, val body: String)
